zh: Really include 'dick' in the name of the license
Previously the translation just said something like 'a man without karma'. This is neither fun nor accurate.

So here is the change -- in Chinese 鸟 means bird literally, and back to the ancient days it is used to refer to 'dick' too. Now this usage has been moved to the character 屌.